+++ src/Genius/Exception/GeniusException.php
@@ -1,9 +1,9 @@
 <?php
+
 declare(strict_types=1);
 
-namespace Genius;
+namespace Genius\Exception;
 
 class GeniusException extends \Exception
 {
-    
 }