+++ src/Genius/Exception/ResourceException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Genius\Exception;
+
+final class ResourceException extends GeniusException
+{
+}