+++ src/Genius/Resources/AccountResource.php
@@ -1,27 +1,20 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Genius\Resources;
 
-use Genius\Authentication\Scope;
+use Genius\Enum\Scope;
+use Genius\Enum\TextFormat;
+use stdClass;
 
 /**
- * Class ArtistsResource
- * @package Genius\Resources
- *
  * @see https://docs.genius.com/#account-h2
  */
-class AccountResource extends AbstractResource
+final class AccountResource extends AbstractResource
 {
-    /**
-     * @param string $text_format
-     * @return \stdClass
-     * @throws ResourceException
-     */
-    public function get(string $text_format = 'dom'): \stdClass
+    public function get(string $text_format = 'dom'): stdClass
     {
-        $this->requireScope(Scope::SCOPE_ME);
-        
-        return $this->sendRequest('GET', 'account/?' . http_build_query(['text_format' => $text_format]));
+        return $this->requester->get('account', ['text_format' => $text_format]);
     }
 }