+++ src/Genius/Resources/SearchResource.php
@@ -1,18 +1,21 @@
 <?php
+
 declare(strict_types=1);
 
 namespace Genius\Resources;
 
+use stdClass;
+
 /**
- * Class ArtistsResource
- * @package Genius\Resources
- *
  * @see https://docs.genius.com/#search-h2
  */
-class SearchResource extends AbstractResource
+final class SearchResource extends AbstractResource
 {
-    public function get(string $query): \stdClass
+    public function get(string $query): stdClass
     {
-        return $this->sendRequest('GET', 'search/?' . http_build_query(['q' => $query]));
+        return $this->requester->get(
+            'search',
+            ['q' => $query]
+        );
     }
 }